+++ tests/JavascriptGeneratorBundle/Fixtures/expected/package-name/src/index.js
@@ -1,0 +1,10 @@
+import { Entity } from '@paysera/http-client-common';
+
+import { createPackageNameClient } from './service/createClient';
+import PackageNameClient from './service/PackageNameClient';
+
+export {
+    Entity,
+    createPackageNameClient,
+    PackageNameClient,
+};