+++ tests/JavascriptGeneratorBundle/Fixtures/expected/package-name/src/service/PackageNameClient.js
@@ -1,0 +1,14 @@
+import { createRequest, ClientWrapper } from '@paysera/http-client-common';
+
+
+class PackageNameClient {
+    /**
+     * @param {ClientWrapper} client
+     */
+    constructor(client) {
+        this.client = client;
+    }
+
+}
+
+export default PackageNameClient;