+++ tests/JavascriptGeneratorBundle/Fixtures/expected/package-version/index.d.ts
@@ -1,0 +1,20 @@
+import { Entity } from '@paysera/http-client-common';
+
+
+interface ClientConfigurationOptions {
+    urlParameters?: {
+        [key: string]: string,
+    },
+    [key: string]: any,
+}
+
+interface ClientConfiguration {
+    baseURL: string,
+    middleware?: object[],
+    options?: ClientConfigurationOptions
+}
+
+export function createPackageVersionClient(configuration: ClientConfiguration): PackageVersionClient;
+
+export interface PackageVersionClient {
+}