+++ tests/PhpGeneratorBundle/RestClient/Fixtures/expected/platform-version/src/ClientFactory.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Paysera\Test\PlatformVersionClient;
+
+use Paysera\Component\RestClientCommon\Util\ClientFactoryAbstract;
+use Paysera\Component\RestClientCommon\Client\ApiClient;
+
+class ClientFactory extends ClientFactoryAbstract
+{
+    const DEFAULT_BASE_URL = 'https://library-address/rest/v1/';
+
+    protected $apiClient;
+
+    public function __construct($options)
+    {
+        if ($options instanceof ApiClient) {
+            $this->apiClient = $options;
+            return;
+        }
+
+        $defaultUrlParameters = [];
+        
+        $options['url_parameters'] = $this->resolveDefaultUrlParameters($defaultUrlParameters, $options);
+        $this->apiClient = $this->createApiClient($options);
+    }
+
+    public function getPlatformVersionClient()
+    {
+        return new PlatformVersionClient($this->apiClient);
+    }
+
+    private function resolveDefaultUrlParameters(array $defaults, array $options)
+    {
+        $params = [];
+        if (isset($options['url_parameters'])) {
+            $params = $options['url_parameters'];
+        }
+
+        return $params + $defaults;
+    }
+}